+++ laktory/models/datasinks/tabledatasink.py
@@ -93,13 +93,6 @@ class TableDataSink(BaseDataSink):
     def _id(self) -> str:
         return self.full_name
 
-    #
-    # @property
-    # def _checkpoint_location(self):
-    #     if self.checkpoint_location:
-    #         return self.checkpoint_location
-    #     return None
-
     # ----------------------------------------------------------------------- #
     # Methods                                                                 #
     # ----------------------------------------------------------------------- #
@@ -168,13 +161,7 @@ class TableDataSink(BaseDataSink):
         """
         Delete sink data and checkpoints
         """
-        if self._checkpoint_location:
-            if os.path.exists(self._checkpoint_location):
-                logger.info(
-                    f"Deleting checkpoint at {self._checkpoint_location}",
-                )
-                shutil.rmtree(self._checkpoint_location)
-
+        # Remove Data
         if self.warehouse == "DATABRICKS":
             logger.info(
                 f"Dropping table {self.full_name}",
@@ -184,6 +171,9 @@ class TableDataSink(BaseDataSink):
             raise NotImplementedError(
                 f"Warehouse '{self.warehouse}' is not yet supported."
             )
+
+        # Remove Checkpoint
+        self._purge_checkpoint(spark=spark)
 
     # ----------------------------------------------------------------------- #
     # Source                                                                  #