+++ laktory/resources/quickstart-stacks/workflows/notebooks/dlt/dlt_laktory_pl.py
@@ -15,7 +15,7 @@ dlt.spark = spark
 logger = get_logger(__name__)
 
 # Read pipeline definition
-pl_name = spark.conf.get("pipeline_name", "dlt-stock-prices")
+pl_name = spark.conf.get("pipeline_name", "pl-stocks-dlt")
 filepath = f"/Workspace{settings.workspace_laktory_root}pipelines/{pl_name}.json"
 with open(filepath, "r") as fp:
     pl = models.Pipeline.model_validate_json(fp.read())
@@ -36,20 +36,43 @@ for udf in pl.udfs:
 # --------------------------------------------------------------------------- #
 
 
-def define_table(node):
+def define_table(node, sink):
+
+    # Get Expectations
+    dlt_warning_expectations = {}
+    dlt_drop_expectations = {}
+    dlt_fail_expectations = {}
+    if not sink.from_quarantine:
+        dlt_warning_expectations = node.dlt_warning_expectations
+        dlt_drop_expectations = node.dlt_drop_expectations
+        dlt_fail_expectations = node.dlt_fail_expectations
+
+    # Get Name
+    name = node.name
+    if sink is not None:
+        name = sink.table_name
+
     @dlt.table_or_view(
-        name=node.name,
+        name=name,
         comment=node.description,
-        as_view=node.sink is None,
+        as_view=sink is None,
     )
-    @dlt.expect_all(node.dlt_warning_expectations)
-    @dlt.expect_all_or_drop(node.dlt_drop_expectations)
-    @dlt.expect_all_or_fail(node.dlt_fail_expectations)
+    @dlt.expect_all(dlt_warning_expectations)
+    @dlt.expect_all_or_drop(dlt_drop_expectations)
+    @dlt.expect_all_or_fail(dlt_fail_expectations)
     def get_df():
-        logger.info(f"Building {node.name} node")
+
+        sink_str = ""
+        if sink is not None:
+            sink_str = f" | sink: {sink.full_name}"
+        logger.info(f"Building {node.name} node{sink_str}")
 
         # Execute node
-        df = node.execute(spark=spark, udfs=udfs)
+        node.execute(spark=spark, udfs=udfs)
+        if sink.from_quarantine:
+            df = node.quarantine_df
+        else:
+            df = node.output_df
         df.printSchema()
 
         # Return
@@ -87,8 +110,15 @@ for node in pl.nodes:
     if node.is_from_cdc:
         df = define_cdc_table(node)
         display(df)
+        continue
 
-    else:
-        wrapper = define_table(node)
+    if node.sinks is None:
+        wrapper = define_table(node, None)
         df = dlt.get_df(wrapper)
         display(df)
+
+    else:
+        for sink in node.sinks:
+            wrapper = define_table(node, sink)
+            df = dlt.get_df(wrapper)
+            display(df)