+++ laktory/resources/quickstart-stacks/workflows/scripts/debug_pl.py
@@ -3,6 +3,7 @@ import importlib
 
 from databricks.connect import DatabricksSession
 from laktory import models
+from laktory import settings
 
 
 # --------------------------------------------------------------------------- #
@@ -10,6 +11,9 @@ from laktory import models
 # --------------------------------------------------------------------------- #
 
 stack_filepath = "../stack.yaml"
+
+# Laktory root on DBFS, required to read checkpoints
+settings.laktory_root = "/laktory/"
 
 spark = DatabricksSession.builder.clusterId("TODO: Fill cluster ID").getOrCreate()